Split IssueSnapshot tests
Moved the GetValue test use TestInitialize into a separate file
